execution/requirements.txt
anthropic>=0.39.0
requests>=2.31.0
python-dotenv>=1.0.0
Pillow>=11.0.0
pilmoji>=2.0.0
